requirements.txt
autogen
openai
python-dotenv

# for faiss index
faiss-cpu
openai
langchain
langdetect
python-docx
PyMuPDF
numpy
python-dotenv
